удаляет моковые предложения
