+++ src/index.tsx
@@ -3,7 +3,6 @@ import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
 import ErrorMessage from './pages/main/components/error-message/error-message';
 import {Provider} from 'react-redux';
-import { offers } from './mocks/offers';
 import { offer } from './mocks/offer';
 import { reviews } from './mocks/reviews';
 import { offersNearby } from './mocks/offers-nearby';
@@ -21,7 +20,7 @@ root.render(
   <React.StrictMode>
     <Provider store={store}>
       <ErrorMessage />
-      <App offers={offers} reviews={reviews} offer={offer} offersNearby={offersNearby} />
+      <App reviews={reviews} offer={offer} offersNearby={offersNearby} />
     </Provider>
   </React.StrictMode>
 );
