--- src/mocks/offers.ts
@@ -1,465 +1,0 @@
-import { City } from '../components/cities-list/cities-list';
-
-export type CityLocation = {
-  'name': City;
-      'location': Location;
-}
-
-export type Location = {
-  'latitude': number;
-  'longitude': number;
-  'zoom': number;
-}
-
-export type OfferData = {
-    'id': string;
-    'title': string;
-    'type': string;
-    'price': number;
-    'previewImage': string;
-    'city': CityLocation;
-    'location': Location;
-    'isFavorite': boolean;
-    'isPremium': boolean;
-    'rating': number;
-  }
-
-export const offers: OfferData[] = [
-  {
-    'id': '70faa463-6bf2-40ec-bca9-4c5d055c5c7f',
-    'title': 'Waterfront with extraordinary view',
-    'type': 'hotel',
-    'price': 217,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/19.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.868610000000004,
-      'longitude': 2.342499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.5,
-  },
-  {
-    'id': 'bfb258c9-01a8-41ca-8261-a89143fc6013',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'room',
-    'price': 197,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/3.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.858610000000006,
-      'longitude': 2.330499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.2,
-  },
-  {
-    'id': 'e721c39a-a261-4481-a9d1-1eca10f9a8a7',
-    'title': 'Waterfront with extraordinary view',
-    'type': 'hotel',
-    'price': 200,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/11.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.834610000000005,
-      'longitude': 2.335499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.3,
-  },
-  {
-    'id': 'f407cc1e-54e1-402a-983f-b9ef5fae611a',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'house',
-    'price': 141,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/18.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.85761,
-      'longitude': 2.358499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 3.1,
-  },
-  {
-    'id': '0fc0b7ce-438c-4824-a4bc-c6ff7b8fc4d4',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'hotel',
-    'price': 491,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/9.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.87561,
-      'longitude': 2.375499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.3,
-  },
-  {
-    'id': '961df615-f3e6-457a-b288-1acab43565bd',
-    'title': 'Beautiful & luxurious apartment at great location',
-    'type': 'hotel',
-    'price': 380,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/4.jpg',
-    'city': {
-      'name': 'Paris',
-      'location': {
-        'latitude': 48.85661,
-        'longitude': 2.351499,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 48.87961000000001,
-      'longitude': 2.353499,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.9,
-  },
-  {
-    'id': 'b6a283d3-0e2d-4039-a5c9-607b415cd56e',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'room',
-    'price': 282,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/2.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.214402,
-      'longitude': 6.764314000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 1.4
-  },
-  {
-    'id': '03d0b121-7feb-48e4-8152-7f22a3a2ccfd',
-    'title': 'Loft Studio in the Central Area',
-    'type': 'hotel',
-    'price': 228,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Dusseldorf',
-      'location': {
-        'latitude': 51.225402,
-        'longitude': 6.776314,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 51.225402,
-      'longitude': 6.784314,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.6
-  },
-  {
-    'id': '453ead1b-e49c-4f58-926d-15006e9b39d5',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'hotel',
-    'price': 450,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/1.jpg',
-    'city': {
-      'name': 'Cologne',
-      'location': {
-        'latitude': 50.938361,
-        'longitude': 6.959974,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.950361,
-      'longitude': 6.961974,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.4
-  },
-  {
-    'id': '99aea0ae-6ec9-4907-87c1-0268e6397d5b',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'room',
-    'price': 166,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Brussels',
-      'location': {
-        'latitude': 50.846557,
-        'longitude': 4.351697,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 50.867557,
-      'longitude': 4.371696999999999,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.1
-  },
-  {
-    'id': 'aba8c1bf-b2e6-44a1-8f32-306d589cc18d',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'room',
-    'price': 286,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/10.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.36554,
-      'longitude': 4.911976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.9
-  },
-  {
-    'id': '1577b10c-037a-40f7-b255-6dd4bfe5ddd1',
-    'title': 'House in countryside',
-    'type': 'apartment',
-    'price': 340,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/13.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.385540000000006,
-      'longitude': 4.902976,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.1
-  },
-  {
-    'id': 'd9295b2a-221e-463d-a250-e91c2f42c89e',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'house',
-    'price': 111,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/20.jpg',
-    'city': {
-      'name': 'Amsterdam',
-      'location': {
-        'latitude': 52.37454,
-        'longitude': 4.897976,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 52.397540000000006,
-      'longitude': 4.9099759999999995,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 4.9
-  },
-  {
-    'id': '61a1576e-b955-41d2-b654-bfc7922bd5ae',
-    'title': 'The Pondhouse - A Magical Place',
-    'type': 'apartment',
-    'price': 271,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/15.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.546341000000005,
-      'longitude': 10.022654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.9
-  },
-  {
-    'id': 'c24db68c-4785-46ca-ae6a-a491d99023f5',
-    'title': 'Nice, cozy, warm big bed apartment',
-    'type': 'house',
-    'price': 472,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.565341000000004,
-      'longitude': 9.978654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': true,
-    'rating': 2.6
-  },
-  {
-    'id': 'd4ebca77-22ed-4288-be73-e1d87225b33d',
-    'title': 'Wood and stone place',
-    'type': 'hotel',
-    'price': 281,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/2.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.570341000000006,
-      'longitude': 9.975654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 3.8
-  },
-  {
-    'id': '705a6c46-f99e-4154-9504-f5076f9ffd46',
-    'title': 'Penthouse, 4-5 rooms + 5 balconies',
-    'type': 'room',
-    'price': 114,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/8.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.558341000000006,
-      'longitude': 10.001654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 4.7
-  },
-  {
-    'id': '010bcbea-1e78-4c9c-88b7-5671dc063e50',
-    'title': 'House in countryside',
-    'type': 'hotel',
-    'price': 418,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/6.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.573341000000006,
-      'longitude': 10.009654000000001,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 2.3
-  },
-  {
-    'id': 'a8c2363f-acf6-4602-9f5f-141dc6b41d3f',
-    'title': 'Tile House',
-    'type': 'room',
-    'price': 273,
-    'previewImage': 'https://15.design.htmlacademy.pro/static/hotel/11.jpg',
-    'city': {
-      'name': 'Hamburg',
-      'location': {
-        'latitude': 53.550341,
-        'longitude': 10.000654,
-        'zoom': 13
-      }
-    },
-    'location': {
-      'latitude': 53.573341000000006,
-      'longitude': 9.994654,
-      'zoom': 16
-    },
-    'isFavorite': false,
-    'isPremium': false,
-    'rating': 1.9
-  },
-];
