+++ src/store/reducer.ts
@@ -2,7 +2,7 @@ import {createReducer} from '@reduxjs/toolkit';
 import {setCity, setSorting, setOffers, requireAuthorization, setOffersDataLoadingStatus, setError} from './action';
 import { City } from '../components/cities-list/cities-list';
 import { Sorting } from '../pages/main/main';
-import { OfferData } from '../mocks/offers';
+import { OfferData } from '../types/offers';
 import { AuthorizationStatus } from '../utils/constants';
 
 type InitialState = {
